Finished most of the middle section of the web page
I added images for the benefits of joining the Gentlemen's Book Club as well as text under the images, and a testimony. I need to fix the height of the first image so everything is proportionate as well as find a way to make the benefit images and text pop out more.