Finished 3/4 sections of the webpage
I am currently working on the bottom of the web page where I am fixing the space issue between the blue box and the footer. I managed to get all sections of the page  to appear nearly proportional  by using width : 30%; for sections 1 and 2 as well as width: 20%; for sections 3 and 4.